# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de guias, eventos e personagens.",
        "title": "Feel The Night API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/char": {
            "post": {
                "description": "Cria um novo personagem com os dados informados do body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Criar um novo personagem",
                "parameters": [
                    {
                        "description": "Dados do Personagem",
                        "name": "character",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCharacterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Cria um usuário com os dados informados no body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Criar um novo usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Buscar usuário por ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.UserErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Deletar usuário por ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.UserErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Atualizar dados do usuário",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.UserErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.UserErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.UserErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateCharacterInput": {
            "type": "object",
            "required": [
                "description",
                "name",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password"
            ],
            "properties": {
                "discordID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateUser": {
            "type": "object",
            "properties": {
                "discordID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UserErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "usuário não encontrado"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "discordID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "eventCount": {
                    "type": "integer"
                },
                "guideCount": {
                    "type": "integer"
                },
                "isModerator": {
                    "type": "boolean"
                },
                "isVeteran": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.Character": {
            "type": "object",
            "properties": {
                "characterID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "guides": {
                    "description": "Has many guides",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Guide"
                    }
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "bannerURL": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "eventID": {
                    "type": "integer"
                },
                "lastEvents": {
                    "description": "Has many LastEvents",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LastEvent"
                    }
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "description": "Foreign Key (Belongs to User)",
                    "type": "integer"
                }
            }
        },
        "models.Guide": {
            "type": "object",
            "properties": {
                "bannerURL": {
                    "type": "string"
                },
                "characterID": {
                    "type": "integer"
                },
                "creationDate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dislikes": {
                    "type": "integer"
                },
                "guideID": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "userID": {
                    "description": "Foreign Keys (Belongs to User and Character)",
                    "type": "integer"
                }
            }
        },
        "models.LastEvent": {
            "type": "object",
            "properties": {
                "eventID": {
                    "description": "Foreign Key (Belongs to Event)",
                    "type": "integer"
                },
                "lastEventID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "discordID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "eventCount": {
                    "type": "integer"
                },
                "events": {
                    "description": "Has many Events and Guides",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                },
                "guideCount": {
                    "type": "integer"
                },
                "guides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Guide"
                    }
                },
                "isModerator": {
                    "type": "boolean"
                },
                "isVeteran": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    }
}